+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - completion for OAuth2 client authentication methods `client_secret_jwt` and `private_key_jwt` [#113](https://github.com/avenga/couper-vscode/pull/113)
+- completion for `beta_roles_map_file` and `beta_permissions_map_file` attributes [#114](https://github.com/avenga/couper-vscode/pull/114)
 
 ---